+++ src/composio_client/types/auth_config_list_response.py
@@ -59,7 +59,6 @@ class Item(BaseModel):
         Literal[
             "OAUTH2",
             "OAUTH1",
-            "OAUTH1A",
             "API_KEY",
             "BASIC",
             "BILLCOM_AUTH",